net/ib, JNI interface: Recv pool size int -> long

+++ src/de/hhu/bsinfo/net/ib/IBConnectionManager.java
@@ -84,7 +84,7 @@ public class IBConnectionManager extends AbstractConnectionManager
         // can't call this in the constructor because it relies on the implemented interfaces for callbacks
         if (!JNIIbdxnet
                 .init(m_coreConfig.getOwnNodeId(), (int) m_config.getIncomingBufferSize().getBytes(), (int) m_config.getOugoingRingBufferSize().getBytes(),
-                        (int) m_config.getIncomingBufferPoolTotalSize().getBytes(), m_config.getMaxRecvReqs(), m_config.getMaxSendReqs(),
+                        m_config.getIncomingBufferPoolTotalSize().getBytes(), m_config.getMaxRecvReqs(), m_config.getMaxSendReqs(),
                         m_config.getFlowControlMaxRecvReqs(), m_config.getMaxConnections(), this, this, this, this, m_config.getEnableSignalHandler(),
                         m_config.getEnableDebugThread())) {
             // #if LOGGER >= DEBUG
